release/projects/reader: componentization and page control
Added a page control component that allows the user to scroll through
the pages of the PDF being displayed.  Also componentized the file
picker widget and removed it from the pdf component.  Finally, some
small changes made to the existing PDF component and widget to enable
display of different pages.

Change-Id: Ia2bb067ef770a607958f2f1b244659fe1d11ff95