TBR reader: adds CSS preprocessing and locally stored PDFs
Adds build tools to preprocess CSS via a browserify transform. This allows CSS
to be created alongside it's corresponding component modules for state creation
and rendering. The pre-processor is based on the one in $V23_ROOT/www and a few
of the basic styles have been copied over.

PDF files can now be stored locally so that they persist between page reloads. A
router was added (based on $V23_ROOT/release/projects/playground/client) to help
support this since some state/input needs to come from the url bar and the
history API.

* Renamed browser/index.js to browser/main.js.
* Moved page-control into pdf component since they are naturally coupled.
* Moved file-picker into footer component.
* Added events/click module.

Change-Id: Ia5885f18bba9bc49c2b42eb6366c1935bea8531a