reader: add styling and initial UI
Added styling to main PDF reading page and the PDF list page.  Also some
additional buttons in the PDF reading page, including an "Open" link to
get back to the list of PDFs (a stub for usability probably to be
replaced later) and the "Link" button that will add or remove the reader
as a member of the linked set for reading.  For now, the link button
connects to a channel that toggles the internal state, but does nothing
more.

Change-Id: I6295128cf19eb0355d0c46ab75a3a8f3fb2de1a1