reader: add device management UI
A new component and related changes that add a UI for managing
the order of devices in a reader device set.  These functions
can be tested through the use of the /#!/overlay route, but
is not yet integrated into the reader app logic.  This will
be done once syncbase integration is a bit further along.

Change-Id: Ic3ec20b5187a209d846e6ae2a4264b2b2a487901