reader/android: show active device sets instead of pdf files
This CL changes the PdfChooserActivity to DeviceSetChooserActivity,
which means that the first screen the user sees shows the list of
currently active device sets.

A device set always has a file attached to it, and when the user
clicks on a device set, the app shows the content of the corresponding
PDF file.

Change-Id: I6e840ac267482aee52b0bc2f2997edfef1eefeb5