reader/android: add floating action button for creating a new device set
Adds a floating action button (FAB) to DeviceSetChooser activity.
When the FAB is touched, the user is directed to a file chooser
activity. Once a pdf file is chosen, the app creates the corresponding
vdl objects and put them in the Syncbase tables.

The previously baked-in PDF files are also removed.

Synchronization works between multiple Android devices, and pages are
always linked by default for now.

NOTE: Not yet using the Syncbase blobs. All the devices need to have
the same pdf files on their local storage.

Change-Id: I1bfeae26f58146e1e9818544e07aea5268948260