reader/android: make it easier to use the local version of android-lib
Recently, there were a few cases where I spent much time debugging
something that has been fixed in the latest go code but not yet
reflected in the published version of io.v:vanadium-android library.

This CL updates the Reader app gradle script so that it can be easily
built with the local version of vanadium-android library by passing
"-DuseLocalVanadiumLib" to the gradle command.

This also updates the vanadium-android library to v0.5, and conforms
to the API changes.

Change-Id: I33ff2756f2b12dd1a45f1114b7f2d00429aa51ec