reader/android: request external storage access permission on Android M+
The log creation was failing on Android M+ because the external
storage permission was not properly granted by the user.
(cf.http://developer.android.com/training/permissions/requesting.html)

This CL explicitly requests the permission, and reinitialize the
loggers once the permission is granted. The logcat code was moved to a
separate DebugUtils class, which could be pushed to the baku-toolkit
level in the future.

Change-Id: Iffc325330f8815d31649e6a8fb62969a7b7f479f