reader/android: adds proof of concept UI tests
Adds a rough, proof of concept version of some automated UI tests. The test
added here will bless the reader application and return to the device list view.
Additional tests and refinement are expected so that we can better duplicate
user scenarios and recreate configurations we know to be problematic so that
debugging can be simplified.

Change-Id: I40e0c32235be3ac314f1cbc52ce29b358c3c9696