reader/android: determine the initial page automatically.
When a device joins an existing device set, it determines the initial
page to show based on what the other devices in the device set are
currently looking at. More specifically, the initial page becomes the
last page of the first consecutive set of pages + 1.

Ex1) When there is only one device looking at page 2: the initial page
of the second device is determined 3.

Ex2) Two devices are looking at pages 3 and 4: the initial page of the
third device is 5.

Ex3) Two devices are looking at pages 3 and 5: the initial page of the
third device is 4.

Closes #53.

Change-Id: Ib52bff8efb37cd7798200df1221dbd1b8c91d5a1