reader/android: fix the problem where device set list is not updated in the initial screen.
When the app launches for the first time, the device set list was not
being updated, because the adapter was being created before the DB is
fully initialized, in which case the adapter fails to monitor the
watch changes coming from Syncbase.

This is fixed by introducing onInitialized() method in DB, which
returns a ListenableFuture which completes when the initialization is
fully performed.

Change-Id: I48d3e5376bc1c834ef37bcb51d041832081dbe43